Use PyNVML 11 for now
PyNVML 12 will depend on nvidia-ml-py and drop its equivalent
functionality. So it will largely become a metapackage with some
utilities. This will mean using a much newer version of the NVML
functionality through Python than is currently used in RAPIDS. As a
preventive step, stick to PyNVML 11, which is what RAPIDS already uses.

Then when PyNVML 12 is released, this can be relaxed on a
project-by-project basis to see if it unconvers any issues that will
need project-level fixes.