Use conda-build instead of conda-mambabuild (#218)
This changes from `conda mambabuild` to `conda build`. Conda now uses
the mamba solver so no performance regressions are expected.

This is a temporary change as we plan to migrate to `rattler-build` in
the near future. However, this is needed sooner to drop `boa` and
unblock Python 3.13 migrations.

xref: https://github.com/rapidsai/build-planning/issues/149