use 'rapids-init-pip' in wheel CI, other CI changes (#235)
## Description

Proposes a batch of small, mostly-unrelated changes to building and
packaging.

These are all part of RAPIDS-wide efforts to make it easier to reproduce
CI locally and to use artifacts from one project's CI in another
project's CI.

Contributes to https://github.com/rapidsai/shared-workflows/issues/356

* explicitly provides an input for `script` to workflows using it,
instead of relying on a default value coming from the workflow file

Contributes to https://github.com/rapidsai/build-planning/issues/179

* adds a call to `rapids-init-pip` near the beginning of all CI scripts
that use `pip`

Also updates `pre-commit` config to the latest
`rapids-dependency-file-generator` version. It was a few versions
behind, thought it'd be ok to bundle that update in with the rest of
these changes.